// src/app/modules/shared/model/image-assets.model.ts
export interface SidebarIconsModel {
  aboutMe: string;
  workExperience: string;
  education: string;
  achievement: string;
  blog: string;
  home: string;
}

export interface SkillsLogoModel {
  angular: string;
  nodeJs: string;
  javaScript: string;
  html: string;
  scss: string;
  css: string;
  firebase: string;
}

export interface SocialLogoModel {
  gmail: string;
  linkedIn: string;
  github: string;
}

export interface ProfileImageModel {
  home: string;
}